Added Dialog Service which allows to display dialogs

+++ App/ViewModels/DiagnoseViewModel.cs
@@ -56,6 +56,8 @@ public class DiagnoseViewModel : ObservableRecipient, INavigationAware
         get;
     }
 
+    private readonly IDialogService _dialogService;
+
     public ICommand DeleteDiagnoseCommand
     {
         get;
@@ -67,13 +69,15 @@ public class DiagnoseViewModel : ObservableRecipient, INavigationAware
         get; 
     }
 
-    public DiagnoseViewModel(ICrudService<Diagnose> crudService, IInfoBarService infoBarService)
+    public DiagnoseViewModel(ICrudService<Diagnose> crudService, IInfoBarService infoBarService, IDialogService dialogService)
     {
         _crudService = crudService;
         InfoBarService = infoBarService;
+        _dialogService = dialogService;
         CreateDiagnoseCommand = new RelayCommand(CreateDiagnose);
         DeleteDiagnoseCommand = new RelayCommand(DeleteDiagnose);
         Diagnoses = new ObservableCollection<Diagnose>();
+        _dialogService = dialogService;
     }
 
 
@@ -87,10 +91,15 @@ public class DiagnoseViewModel : ObservableRecipient, INavigationAware
     }
     private async void DeleteDiagnose()
     {
-        Diagnose diagnoseToRemove = SelectedItem;
-        Diagnoses.Remove(diagnoseToRemove);
-        await _crudService.Delete(diagnoseToRemove);
-        InfoBarService.showMessage("Erfolgreich gelöscht","Erfolgreich");
+        var result = await _dialogService.ConfirmDeleteDialogAsync("Fehlerkategorie Löschen", "Sind Sie sicher das Sie den Eintrag löschen wollen?");
+        if (result != null && result == true)
+        {
+            Diagnose diagnoseToRemove = SelectedItem;
+            Diagnoses.Remove(diagnoseToRemove);
+            await _crudService.Delete(diagnoseToRemove);
+            InfoBarService.showMessage("Erfolgreich gelöscht", "Erfolgreich");
+        }
+       
     }
 
     public async void OnNavigatedTo(object parameter)
